backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
python-multipart==0.0.9
Pillow==10.4.0
numpy==1.26.4
google-api-python-client==2.151.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
reportlab==4.2.5

